+++ app/src/services/ScoreService.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 
 export const updateScores = async (userID: number) => {
   try {
-    await axios.post(`http://localhost:8080/recipes/score?userID=${userID}`);
+    await axios.post(`http://localhost:8080/score?userID=${userID}`);
   } catch (error) {
     console.error("Failed to fetch search query", error);
   }
